pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "worth-my-time"
version = "0.1.0"
description = "Local-first content triage: Brave bookmarks -> Markdown analyses (anti-clickbait, FOMO-friendly)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0", "youtube-transcript-api>=0.6.0"]

[project.scripts]
wmt = "wmt.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
